pyproject.toml
[tool.poetry]
name = "bundler-spec-tests"
version = "0.1.0"
description = ""
authors = ["shahafn <shahaflol@gmail.com>"]
readme = "README.md"
packages = [{include = "bundler_spec_tests"}]

[tool.poetry.dependencies]
python = "^3.8"



[tool.pdm]
[tool.pdm.build]
includes = ["bundler_spec_tests"]
[build-system]
requires = ["pdm-pep517>=1.0.0"]
build-backend = "pdm.pep517.api"

[tool.pdm.scripts]
init-deps = {shell = "cd account-abstraction; yarn"}
deploy = {shell = "cd account-abstraction; yarn deploy --network localhost"}

[project]
name = "bundler-spec-tests"
version = "0.1.0"
description = ""
authors = [
    {name = "shahafn", email = "shahaflol@gmail.com"},
]
dependencies = [
    "pytest>=7.2.0",
    "requests>=2.28.1",
    "setuptools>=65.6.0",
    "eth-utils>=1.2.0",
    "py-solc-x>=1.1.1",
    "eth-tester>=0.6.0b6",
    "web3>=5.31.1",
]
requires-python = ">=3.8,<4.0"
license = {text = "MIT"}
readme = "README.md"

[project.urls]
